[SPARK-8890][SQL] Fallback on sorting when writing many dynamic partitions
